Fix `Linspace` integer-input dtype promotion: int → FLOAT64, not FLOAT32.
The Javadoc comment "tf.linspace promotes integer start/stop to float32"
and the matching arm in `getDefaultDTypes` were both wrong. TF actually
promotes integer start/stop to FLOAT64. Verified empirically on TF 2.9:

    >>> tf.linspace(tf.constant(0, dtype=tf.int32),
    ...             tf.constant(10, dtype=tf.int32), 5).dtype
    <dtype: 'float64'>

The bug was masked by the float-input fixture (`tf2_test_linspace.py`,
which uses `0.0`/`1.0` and so never reaches the int-promotion arm); a
follow-up commit adds an int-input fixture that exercises and asserts
the corrected behavior.

Surfaced when investigating the codecov-flagged partial coverage on
`Linspace.getDefaultDTypes` — the uncovered branch was hiding the bug.

+++ com.ibm.wala.cast.python.ml/source/com/ibm/wala/cast/python/ml/client/Linspace.java
@@ -18,11 +18,12 @@ import java.util.Set;
 
 /**
  * Generator for {@code tf.linspace(start, stop, num, name=None, axis=0)}. Output is a 1-D tensor of
- * length {@code num}; output dtype follows {@code start} (with int → float32 promotion per TF
- * semantics: integer start/stop produce a float32 result). The {@code axis} parameter is honored
- * only at its default value of 0 (the rank-1 case); non-default axes return ⊤ shape since they
- * require start/stop to be tensors and the result-shape derivation depends on broadcasting that
- * isn't tracked here.
+ * length {@code num}; output dtype follows {@code start} (with int → float64 promotion per TF
+ * semantics — verified empirically on TF 2.9: {@code tf.linspace(tf.constant(0, dtype=tf.int32),
+ * tf.constant(10, dtype=tf.int32), 5).dtype} is {@code float64}, not {@code float32}). The {@code
+ * axis} parameter is honored only at its default value of 0 (the rank-1 case); non-default axes
+ * return ⊤ shape since they require start/stop to be tensors and the result-shape derivation
+ * depends on broadcasting that isn't tracked here.
  *
  * @see <a href="https://www.tensorflow.org/api_docs/python/tf/linspace">tf.linspace</a>
  * @author <a href="mailto:khatchad@hunter.cuny.edu">Raffi Khatchadourian</a>
@@ -115,8 +116,8 @@ public class Linspace extends TensorGenerator {
     if (startDTypes == null || startDTypes.isEmpty()) return EnumSet.of(DType.UNKNOWN);
     Set<DType> ret = new HashSet<>();
     for (DType dt : startDTypes) {
-      // tf.linspace promotes integer start/stop to float32.
-      if (dt == DType.INT32 || dt == DType.INT64) ret.add(DType.FLOAT32);
+      // tf.linspace promotes integer start/stop to float64 (verified on TF 2.9).
+      if (dt == DType.INT32 || dt == DType.INT64) ret.add(DType.FLOAT64);
       else ret.add(dt);
     }
     return ret;
